17. Data Validation

+++ Finshark/Controllers/CommentController.cs
@@ -23,15 +23,25 @@ namespace Finshark.Controllers
         [Route("GetAll")]
         public async Task<IActionResult> GetAll()
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var comments = await _commentRepository.GetAllAsync();
             var commentDto = comments.Select(s => s.ToCommentDto());
             return Ok(commentDto);
         }
 
         [HttpGet]
-        [Route("GetById/{id}")]
+        [Route("GetById/{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var comment = await _commentRepository.GetByIdAsync(id);
             if(comment == null)
             {
@@ -41,10 +51,15 @@ namespace Finshark.Controllers
         }
 
         [HttpPost]
-        [Route("Create/{stockId}")]
+        [Route("Create/{stockId:int}")]
         public async Task<IActionResult> Create([FromRoute] int stockId, [FromBody] CreateCommentDto commentDto)
         {
-            if(!await _stockRepository.StockExists(stockId))
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
+            if (!await _stockRepository.StockExists(stockId))
             {
                 return BadRequest("Stock does not exist");
             }
@@ -56,9 +71,14 @@ namespace Finshark.Controllers
         }
 
         [HttpPut]
-        [Route("update/{id}")]
+        [Route("update/{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentRequestDto updateDto)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var comment = await _commentRepository.UpdateAsync(id, updateDto.ToCommentFromUpdate(id));
             if (comment == null)
             {
@@ -69,9 +89,14 @@ namespace Finshark.Controllers
 
 
         [HttpDelete]
-        [Route("delete/{id}")]
+        [Route("delete/{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var commentModel = await _commentRepository.DeleteAsync(id);
             if(commentModel == null)
             {
