+++ Finshark/Controllers/StockController.cs
@@ -23,6 +23,11 @@ namespace Finshark.Controllers
         [Route("GetAll")]
         public async Task<IActionResult> GetAll()
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var stocks = await _stockRepository.GetAllAsync(); 
             var stockDto = stocks.Select(s => s.ToStockDto());    
 
@@ -31,9 +36,14 @@ namespace Finshark.Controllers
 
 
         [HttpGet]
-        [Route("GetById/{id}")]
+        [Route("GetById/{id:int}")]
         public async Task<IActionResult> GetById([FromRoute]  int id) 
-        { 
+        {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var stock =  await _stockRepository.GetByIdAsync(id);
 
             if (stock == null)
@@ -49,6 +59,11 @@ namespace Finshark.Controllers
         [Route("Create")]
         public async Task<IActionResult> Create([FromBody] CreateStockRequestDto stockDto)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var stockModel = stockDto.ToStockFromCreateDto();
             await _stockRepository.CreateAsync(stockModel);
 
@@ -57,9 +72,14 @@ namespace Finshark.Controllers
 
 
         [HttpPut]
-        [Route("Update/{id}")]
+        [Route("Update/{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockRequestDto updateDto )
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var stockModel = await _stockRepository.UpdateAsync(id, updateDto);
             if (stockModel == null)
             {
@@ -71,9 +91,14 @@ namespace Finshark.Controllers
 
 
         [HttpDelete]
-        [Route("Delete/{id}")]
+        [Route("Delete/{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var stockModel = await _stockRepository.DeleteAsync(id);
             if(stockModel == null)
             {
